+++ python/cudf/cudf/core/dataframe.py
@@ -2545,11 +2545,8 @@ class DataFrame(IndexedFrame, Serializable, GetAttrGetItemMixin):
             inplace=inplace,
         )
 
-    def take(self, indices, axis=0, keep_index=None):
-        axis = self._get_axis_from_axis_arg(axis)
-        if axis != 0:
-            raise NotImplementedError("Only axis=0 is supported.")
-        out = super().take(indices, keep_index)
+    def take(self, indices, axis=0):
+        out = super().take(indices)
         out.columns = self.columns
         return out
